+++ Sources/Tools/MaxConvert/hsControlConverter.cpp
@@ -679,8 +679,8 @@ plCompoundController *hsControlConverter::MakeTransformController(Control *contr
 
 void hsControlConverter::ISetSegRange(float start, float end)
 {
-    fSegStart = (start >= 0 ? fTicksPerSec * start : fInterface->GetAnimRange().Start());
-    fSegEnd = (end >= 0 ? fTicksPerSec * end : fInterface->GetAnimRange().End());
+    fSegStart = (start >= 0 ? int(fTicksPerSec * start) : fInterface->GetAnimRange().Start());
+    fSegEnd = (end >= 0 ? int(fTicksPerSec * end) : fInterface->GetAnimRange().End());
 }
 
 
@@ -1000,7 +1000,7 @@ int hsControlConverter::IAddPartsKeys(Control* control,
         {
             // Get key
             ikeys->GetKey(i, key.get());
-            float frameTime = key->time / GetTicksPerSec();
+            float frameTime = float(key->time) / GetTicksPerSec();
             int frameNum = key->time / GetTicksPerFrame();
             hsAssert(frameNum <= hsKeyFrame::kMaxFrameNumber, "Anim is too long.");
